refactor(editor): utilize ui related logic
